Add process tree tracking to Node class
Process parent and child relationships are now tracked by the Node.
This should allow proper management and cleanup of closed processes.
Process Ids are now treated as int's everywhere to simplify handling.

+++ HackLinks Server/Session.cs
@@ -55,7 +55,7 @@ namespace HackLinks_Server
         }
         */
 
-        public bool HasProcessId(long pid)
+        public bool HasProcessId(int pid)
         {
             return attachedProcess.ProcessId == pid;
         }
